Revision to handle large tables where count_big(*) is required

+++ src/Microsoft.Health.Fhir.Core/Features/Operations/Reindex/Models/ReindexJobRecord.cs
@@ -119,10 +119,10 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Reindex.Models
         public ConcurrentDictionary<ReindexJobQueryStatus, byte> QueryList { get; private set; } = new ConcurrentDictionary<ReindexJobQueryStatus, byte>();
 
         [JsonProperty(JobRecordProperties.ResourceCounts)]
-        public ConcurrentDictionary<string, int> ResourceCounts { get; private set; } = new ConcurrentDictionary<string, int>();
+        public ConcurrentDictionary<string, long> ResourceCounts { get; private set; } = new ConcurrentDictionary<string, long>();
 
         [JsonProperty(JobRecordProperties.Count)]
-        public int Count { get; set; }
+        public long Count { get; set; }
 
         [JsonProperty(JobRecordProperties.Progress)]
         public int Progress { get; set; }
